refactor(webapp): extract app bootstrap into src/app/, clean up src root
- Create src/app/ with App.tsx, providers.tsx, router.tsx
- Move store.ts -> store/index.ts (alongside store/playground.ts)
- Move store.unit.test.ts -> store/index.unit.test.ts
- Add utils/constants.ts for PROD_ENVIRONMENT_NAME (no window side-effects)
- Replace INTEGRATION_TEMPLATES_GITHUB_URL with githubRepo from utils/cloud
- Delete reportWebVitals.ts (CRA no-op boilerplate)
- Delete constants.ts (contents distributed)
- Fix .js-extension store imports that broke after the move
- index.tsx is now a thin createRoot().render() entry point
