--- packages/webapp/src/constants.ts
@@ -1,2 +1,0 @@
-export const PROD_ENVIRONMENT_NAME = 'prod';
-export const INTEGRATION_TEMPLATES_GITHUB_URL = 'https://github.com/NangoHQ/integration-templates';
