+++ packages/webapp/src/index.tsx
@@ -1,16 +1,10 @@
-import { QueryClientProvider } from '@tanstack/react-query';
-import { NuqsAdapter } from 'nuqs/adapters/react-router/v6';
 import posthog from 'posthog-js';
-import { PostHogProvider } from 'posthog-js/react';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from './App';
-import { ErrorBoundary } from './components/ErrorBoundary';
-import reportWebVitals from './reportWebVitals';
-import { queryClient } from './store';
+import App from './app/App';
+import { Providers } from './app/providers';
 import { globalEnv } from './utils/env';
-import { SentryErrorBoundary } from './utils/sentry';
 
 import './index.css';
 
@@ -27,19 +21,8 @@ if (globalEnv.publicPosthogKey) {
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
     <React.StrictMode>
-        <SentryErrorBoundary fallback={<ErrorBoundary />}>
-            <PostHogProvider client={posthog}>
-                <NuqsAdapter>
-                    <QueryClientProvider client={queryClient}>
-                        <App />
-                    </QueryClientProvider>
-                </NuqsAdapter>
-            </PostHogProvider>
-        </SentryErrorBoundary>
+        <Providers>
+            <App />
+        </Providers>
     </React.StrictMode>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
