+++ packages/webapp/src/pages/Environment/Settings/General.tsx
@@ -7,7 +7,7 @@ import { permissions } from '@nangohq/authz';
 import { DeleteButton } from './components/DeleteButton';
 import SettingsContent from './components/SettingsContent';
 import SettingsGroup from './components/SettingsGroup';
-import { PROD_ENVIRONMENT_NAME } from '../../../constants';
+import { PROD_ENVIRONMENT_NAME } from '../../../utils/constants';
 import { useDeleteEnvironment, useEnvironment, usePatchEnvironment } from '../../../hooks/useEnvironment';
 import { useMeta } from '../../../hooks/useMeta';
 import { useStore } from '../../../store';
