+++ packages/webapp/src/pages/Integrations/CreateList.tsx
@@ -9,7 +9,7 @@ import { useNavigate } from 'react-router-dom';
 import { AuthBadge } from './components/AuthBadge.js';
 import { useProviders } from '../../hooks/useProviders.js';
 import DashboardLayout from '../../layout/DashboardLayout.js';
-import { useStore } from '../../store.js';
+import { useStore } from '@/store';
 import { IntegrationLogo } from '@/components-v2/IntegrationLogo';
 import { Badge } from '@/components-v2/ui/badge';
 import { InputGroup, InputGroupAddon, InputGroupInput } from '@/components-v2/ui/input-group';
