+++ packages/webapp/src/pages/Integrations/components/FunctionSwitch.tsx
@@ -5,7 +5,7 @@ import { permissions } from '@nangohq/authz';
 import { useEnvironment } from '../../../hooks/useEnvironment.js';
 import { useFlowDisable, useFlowEnable, usePreBuiltDeployFlow } from '../../../hooks/useFlow.js';
 import { useToast } from '../../../hooks/useToast.js';
-import { useStore } from '../../../store.js';
+import { useStore } from '@/store';
 import { APIError } from '../../../utils/api.js';
 import { PermissionGate } from '@/components-v2/PermissionGate';
 import { Switch } from '@/components-v2/ui/switch';
