+++ packages/webapp/src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const PROD_ENVIRONMENT_NAME = 'prod';
